Add files via upload
Virtual Desktop Assistant:
This is the project whose goal is to learn machine learning and artificial intelligence technology by implementing python script in This (Virtual Assistant) project.

The main task of a virtual assistant is to minimize the use of input devices like keyboard, mouse, touch pens. This is more convenient way to give input command through the voice to the computer instead giving the command through the keyboard. It will become more comfortable to the user who actually don't know how to do the particular task physically. It will customize their work by letting do the task by system itself only using user's voice command.
The UI of this project make the virtual assistant user friendly so that they can easily interact with it.

The project is done in python using python libraries and modules and the good looking UI is done using pyQt5 which is the most powerful GUI library in python.